moving bert to semantic metrics
