+++ continuous_eval/metrics/__init__.py
@@ -1,9 +1,11 @@
 from continuous_eval.metrics.base import Metric
 from continuous_eval.metrics.generation_deterministic_metrics import (
-    BertAnswerRelevance,
-    BertAnswerSimilarity,
     DeterministicAnswerRelevance,
     DeterministicFaithfulness,
+)
+from continuous_eval.metrics.generation_semantic_metrics import (
+    BertAnswerRelevance,
+    BertAnswerSimilarity,
 )
 from continuous_eval.metrics.generation_LLM_based_metrics import (
     LLMBasedAnswerCorrectness,
